+++ test/OptionT.ts
@@ -1,22 +1,22 @@
 import * as assert from 'assert'
 import * as O from '../src/Option'
 import { getOptionM } from '../src/OptionT'
-import { task } from '../src/Task'
+import * as T from '../src/Task'
 
-const T = getOptionM(task)
+const MT = getOptionM(T.monadTask)
 
 describe('OptionT', () => {
   it('map', () => {
-    const greetingT = T.of('welcome')
-    const excitedGreetingT = T.map(greetingT, (s) => s + '!')
+    const greetingT = MT.of('welcome')
+    const excitedGreetingT = MT.map(greetingT, (s) => s + '!')
     return excitedGreetingT().then((o) => {
       assert.deepStrictEqual(o, O.some('welcome!'))
     })
   })
 
   it('chain', () => {
-    const to1 = T.chain(T.of('foo'), (a) => T.of(a.length))
-    const to2 = T.chain(task.of(O.none), (a: string) => T.of(a.length))
+    const to1 = MT.chain(MT.of('foo'), (a) => MT.of(a.length))
+    const to2 = MT.chain(T.of(O.none), (a: string) => MT.of(a.length))
     return Promise.all([to1(), to2()]).then(([o1, o2]) => {
       assert.deepStrictEqual(o1, O.some(3))
       assert.deepStrictEqual(o2, O.none)
@@ -24,36 +24,36 @@ describe('OptionT', () => {
   })
 
   it('fold', async () => {
-    const f = () => task.of('none')
-    const g = (s: string) => task.of(`some${s.length}`)
-    const s1 = await T.fold(task.of(O.none), f, g)()
+    const f = () => T.of('none')
+    const g = (s: string) => T.of(`some${s.length}`)
+    const s1 = await MT.fold(T.of(O.none), f, g)()
     assert.deepStrictEqual(s1, 'none')
-    const s2 = await T.fold(T.of('s'), f, g)()
+    const s2 = await MT.fold(MT.of('s'), f, g)()
     assert.deepStrictEqual(s2, 'some1')
   })
 
   it('alt', async () => {
-    const o1 = await T.alt(task.of(O.some(1)), () => task.of(O.some(2)))()
+    const o1 = await MT.alt(T.of(O.some(1)), () => T.of(O.some(2)))()
     assert.deepStrictEqual(o1, O.some(1))
-    const o2 = await T.alt(task.of(O.none), () => task.of(O.some(2)))()
+    const o2 = await MT.alt(T.of(O.none), () => T.of(O.some(2)))()
     assert.deepStrictEqual(o2, O.some(2))
   })
 
   it('getOrElse', async () => {
-    const n1 = await T.getOrElse(task.of(O.some(1)), () => task.of(2))()
+    const n1 = await MT.getOrElse(T.of(O.some(1)), () => T.of(2))()
     assert.deepStrictEqual(n1, 1)
-    const n2 = await T.getOrElse(task.of(O.none), () => task.of(2))()
+    const n2 = await MT.getOrElse(T.of(O.none), () => T.of(2))()
     assert.deepStrictEqual(n2, 2)
   })
 
   it('fromM', () => {
-    return T.fromM(task.of(1))().then((o) => {
+    return MT.fromM(T.of(1))().then((o) => {
       assert.deepStrictEqual(o, O.some(1))
     })
   })
 
   it('none', () => {
-    return T.none()().then((o) => {
+    return MT.none()().then((o) => {
       assert.deepStrictEqual(o, O.none)
     })
   })